# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  github_com_merka_api_internal_domain.Product:
    properties:
      ativo:
        type: boolean
      categoryID:
        type: string
      cfop:
        type: string
      id:
        type: string
      ncm:
        description: |-
          NCM/CFOP: obrigatórios para emitir NFC-e via SEFAZ direta (ETAPA 4,
          ver CLAUDE.md) — nil em produtos cadastrados antes da migration
          0014. internal/fiscal.FiscalProviderSefazDireto recusa emitir nota
          com item sem esses dados em vez de estimar um valor.
        type: string
      nome:
        type: string
      precoPorKg:
        description: usado quando TipoCobranca == TipoCobrancaPeso
        format: float64
        type: number
      precoUnitario:
        description: usado quando TipoCobranca == TipoCobrancaUnitario
        format: float64
        type: number
      taraKg:
        description: peso do prato/recipiente, descontado do peso bruto lido na balança
        format: float64
        type: number
      tenantID:
        type: string
      tipoCobranca:
        $ref: '#/definitions/github_com_merka_api_internal_domain.TipoCobranca'
    type: object
...
  github_com_merka_api_internal_domain.TipoCobranca:
    enum:
    - unitario
    - peso
    type: string
    x-enum-varnames:
    - TipoCobrancaUnitario
    - TipoCobrancaPeso
...
  internal_handler.cadastrarProdutoRequest:
    properties:
      category_id:
        type: string
      nome:
        type: string
      preco_por_kg:
        type: number
      preco_unitario:
        type: number
      tara_kg:
        type: number
      tipo_cobranca:
        type: string
    type: object
...
  internal_handler.configurarPrecoPesoRequest:
    properties:
      preco_por_kg:
        type: number
      tara_kg:
        type: number
    type: object
...
info:
  contact: {}
  description: Sistema de comandas multi-tenant (churrascaria como primeira instância).
    Ver CLAUDE.md e docs/merka-planejamento.md para o contexto completo do domínio.
  title: Merka API
  version: "1.0"
paths:
...
  /produtos:
    get:
      description: Qualquer perfil autenticado — usado por garçom/balança pra escolher
        o que lançar (US-09/US-11).
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_merka_api_internal_domain.Product'
            type: array
        "401":
          description: token ausente, inválido ou expirado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Listar catálogo de produtos ativos
      tags:
      - produtos
    post:
      consumes:
      - application/json
      description: Restrito a Admin Super, Gestor ou Caixa (permissão "cadastrar_produto").
        tipo_cobranca "unitario" exige preco_unitario > 0; tipo_cobranca "peso" exige
        preco_por_kg > 0 (tara_kg opcional, default 0). Grava a primeira linha em
        product_price_history para produtos do tipo peso.
      parameters:
      - description: Dados do produto
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.cadastrarProdutoRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_merka_api_internal_domain.Product'
        "400":
          description: campo obrigatório ausente ou tipo_cobranca inválido
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: token ausente, inválido ou expirado
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: usuário sem permissão para esta ação
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Cadastrar novo produto no catálogo (US-21)
      tags:
      - produtos
  /produtos/{id}/preco-peso:
    patch:
      consumes:
      - application/json
      description: Restrito a Admin Super, Gestor, Caixa ou Balança (permissão "configurar_preco_peso")
        — ajuste operacional do dia a dia, diferente das demais configurações estruturais.
        Só se aplica a produtos do tipo peso. Grava o histórico da alteração em product_price_history.
      parameters:
      - description: ID do produto
        in: path
        name: id
        required: true
        type: string
      - description: Novo preço/kg e/ou tara (ao menos um)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handler.configurarPrecoPesoRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_merka_api_internal_domain.Product'
        "400":
          description: nada para atualizar
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: token ausente, inválido ou expirado
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: usuário sem permissão para esta ação
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: produto não encontrado
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: produto não é do tipo peso
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: erro interno
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Configurar preço/kg e tara de um produto (US-20)
      tags:
      - produtos
...
securityDefinitions:
  BearerAuth:
    description: Informe "Bearer <token>" (token obtido em POST /auth/login)
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"